[WTEL-6107] moved notification error with api to store
