+++ src/modules/configuration/modules/lookups/modules/slas/modules/conditions/store/conditions.js
@@ -4,6 +4,13 @@ import {
   createCardStoreModule,
   createTableStoreModule,
 } from '@webitel/ui-sdk/store';
+import applyTransform, {
+  notify,
+} from '@webitel/ui-sdk/src/api/transformers/index';
+
+import i18n from '../../../../../../../../../app/locale/i18n.js';
+
+const { t } = i18n.global;
 
 import ConditionsAPI from '../api/conditions.js';
 import filters from '../modules/filters/store/filters.js';
@@ -28,7 +35,21 @@ const getters = {
 
 const actions = {
   ADD_ITEM: async (context) => {
-    await context.dispatch('api/POST_ITEM', { context });
+    try {
+      await context.dispatch('api/POST_ITEM', { context });
+    } catch (err) {
+      throw applyTransform(err, [
+        notify(({ callback }) =>
+          callback({
+            type: 'error',
+            text:
+              err.response.data.code === 400
+                ? t('lookups.slas.duplicateConditionName')
+                : err.response?.data?.detail,
+          }),
+        ),
+      ]);
+    }
   },
   UPDATE_ITEM: async (context) => {
     await context.dispatch('api/UPD_ITEM', { context });
